internal: refactor mismatched args count diagnostic
